BS-111 Patient's detail view (examinations table)
